Crate telamon_utils

Generic helper functions.

Re-exports

pub use self::ndarray::NDArray;

pub use self::ndarray::NDRange;

Modules

multimap
A FnvHashMap with mutiple values for each key.

ndarray
An array with a variable number of dimensions.

tfrecord
This module provides Read and Write extensions for reading and writing TFRecord files. TFRecord is a simple container file format for embedding a sequence of data used notably by TensorFlow (see https://www.tensorflow.org/api_guides/python/python_io).

unwrap
A macro to help debug unwraps.

Macros

eq_from_key
Derives Eq based on a method that returns a key for the object.

generated_file
Includes a generates file into the current file.

hash_from_key
Derives Eq and Hash based on a method that returns a key for the object.

unwrap
Panics after if the value cannot be unwraped.

Structs

Cache
A thread-safe LRU Cache.

Dag
A directed acyclic graph given by adjacency list.

FilterList
Iterates over a linked list while removing some items.

PartialPermutations

RcStr
A reference counted string, compatible with &str.

VecSet
A set backed by an ordered vector.

ZipCopy
Zip copies of an object with an iterator.

Enums

Trivalent
Booleans enhanced with a third Maybe value.

Traits

BuilderTrait
A trait that implements useful methods on builders.

Functions

at_most_one
Ensures an iterator has at most one element.

clone_pair
Clones a pair of reference.

cmp_f64
Compare two f64, panic if one of them is nan

div_ceil
Performs an integer divison rounded to the upper number.

filter_list
Iterates over a linked list while removing some items.

log2_u32
Returns the log2 of a power of 2.

to_map
Transforms an iterator into an FnvHashMap. Redundant nodes are merged using merge.

zip_copy
Zip copies of an object with an iterator.

Type Definitions

FnvHashMap
A HashMap using a default FNV hasher.

FnvHashSet
A HashSet using a default FNV hasher.

FnvMultiHashMap