Enum telamon_utils::Trivalent

pub enum Trivalent {
    False,
    Maybe,
    True,
}

Booleans enhanced with a third Maybe value.

Variants
FalseMaybeTrue
Methods

impl Trivalent

pub fn as_bool(&self) -> Option<bool>

Returns the boolean represented.

pub fn maybe_true(&self) -> bool

Returns true if the underlying boolean might be true.

pub fn maybe_false(&self) -> bool

Returns true if the underlying boolean might be false.

pub fn is_true(&self) -> bool

Returns true if the underlying boolean is true.

pub fn is_false(&self) -> bool

Returns true if the underlying boolean is false.

pub fn is_maybe(&self) -> bool

Returns true if the underlying boolean might be true and false.