Struct telamon_utils::multimap::MultiHashMap

pub struct MultiHashMap<K, V, S = RandomState> where
    K: Hash + Eq,
    S: BuildHasher,  { /* fields omitted */ }

A HashMap with mutiple values for each key.

Methods

impl<K: Hash + Eq, V, S: BuildHasher> MultiHashMap<K, V, S>

pub fn new() -> MultiHashMap<K, V, RandomState>

Creates an empty MultiHashMap.

pub fn with_capacity(capacity: usize) -> MultiHashMap<K, V, RandomState>

Creates an empty hash map with the given initial capacity.

pub fn with_hasher(hash_builder: S) -> Self

Creates an empty MultiHashMap which will use the given hash builder to hash keys.

pub fn with_capacity_and_hasher(capacity: usize, hasher: S) -> Self

Creates an empty MultiHashMap with space for at least capacity elements, using hasher to hash the keys.

pub fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the MultiHashMap. The collection may reserve more space to avoid frequent reallocations.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the map as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

Important traits for Keys<'a, K, V>
impl<'a, K, V> Iterator for Keys<'a, K, V> type Item = &'a K;

pub fn keys(&self) -> Keys<K, Vec<V>>

An iterator visiting all keys in arbitrary order.

pub fn values(&self) ->

An iterator visiting all values in arbitrary order.

pub fn values_mut(&mut self) ->

An iterator visitinf all values mutably in arbitrary order.

Important traits for Iter<'a, K, V>
impl<'a, K, V> Iterator for Iter<'a, K, V> type Item = (&'a K, &'a V);

pub fn iter(&self) -> Iter<K, Vec<V>>

Iterates over all the keys and returns the associated values.

pub fn num_keys(&self) -> usize

Returns the number of keys.

pub fn is_empty(&self) -> bool

Indicates if the map contains no elements.

Important traits for Drain<'a, K, V>
impl<'a, K, V> Iterator for Drain<'a, K, V> type Item = (K, V);

pub fn drain(&mut self) -> Drain<K, Vec<V>>

Clears the map, returning all key-value pairs as an iterator. Keeps the allocated memory for reuse.

pub fn clear(&mut self)

Clears the map, removing all keys and values.

pub fn get<'a, Q: ?Sized>(&'a self, k: &Q) -> <'a> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns the values mapped to the key.

pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> bool where
K: Borrow<Q>,
Q: Hash + Eq,

Indicates if the map contains a value for the specified key.

pub fn get_mut<'a, Q: ?Sized>(&'a mut self, k: &Q) -> <'a> where
K: Borrow<Q>,
Q: Hash + Eq,

Returns an iterator over mutable reference to the values mapped to the key.

pub fn insert(&mut self, k: K, v: V)

Inserts new value to the map.

pub fn insert_many(&mut self, k: K, values: Vec<V>)

Inserts new value to the map.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn remove<Q: ?Sized>(&mut self, k: &Q) -> Vec<V> where
K: Borrow<Q>,
Q: Hash + Eq,

Removes all the elements bound to a key.