Trait telamon_utils::tfrecord::RecordReader

pub trait RecordReader: Read {
    fn read_record(&mut self, buf: &mut Vec<u8>) -> Result<usize, ReadError> { ... }

    fn records(self) -> Records<Self>
    where
        Self: Sized,
    { ... }
}

A trait extension for reading records.

Inspired from the C++ implementation at: * https://github.com/tensorflow/tensorflow/blob/f318765ad5a50b2fbd7cc08dd4ebc249b3924270/tensorflow/core/lib/io/record_reader.h

https://github.com/tensorflow/tensorflow/blob/f318765ad5a50b2fbd7cc08dd4ebc249b3924270/tensorflow/core/lib/io/record_reader.cc

Provided methods

fn read_record(&mut self, buf: &mut Vec<u8>) -> Result<usize, ReadError>

Read a single record, placing the bytes into buf.

All bytes read from this source will be appended to the specified buffer buf.

If successful, this function returns the total number of bytes read, including the tfrecord header and footer sizes. If the total number of bytes read is 0, then the reader has reached end of file.

Errors

If this function encounters an error of the kind ErrorKind::Interrupted then the error is ignored and the operation will continue.

If any other read error is encountered then this function immediately returns. Any bytes which have already been read will be appended to buf. If an error occurs while reading the initial tfrecord header, buf is unchanged.

Important traits for Records<R>
impl<R: Read> Iterator for Records<R> type Item = Result<Vec<u8>, ReadError>;

fn records(self) -> Records<Self> where
Self: Sized,

Transforms this Read instance to an Iterator over the contained records.

The returned type implements Iterator where the Item is Result<u8, ReadError>. The yielded item is Ok if a record was successfuly read and Err otherwise. EOF is mapped to returning None from this iterator.

Loading content...