Trait telamon_utils::tfrecord::RecordWriter

pub trait RecordWriter: Write {
    type Writer;
    fn finish(self) -> Result<Self::Writer, WriteError>;

    fn finish_box(self: Box<Self>) -> Result<Self::Writer, WriteError>;

    fn write_record(&mut self, bytes: &[u8]) -> Result<(), WriteError> { ... }
}

A trait extension for writing records.

Inspired from the C++ implementation at: * https://github.com/tensorflow/tensorflow/blob/f318765ad5a50b2fbd7cc08dd4ebc249b3924270/tensorflow/core/lib/io/record_writer.h

https://github.com/tensorflow/tensorflow/blob/f318765ad5a50b2fbd7cc08dd4ebc249b3924270/tensorflow/core/lib/io/record_writer.cc

Associated Types

type Writer

Required Methods

fn finish(self) -> Result<Self::Writer, WriteError>

Writes all output to the file and conclude the stream. Returns the underlying writer. Does not flush the underlying writer.

This is the conceptual equivalent of RecordWriter::Close in TensorFlow's C++ implementation.

fn finish_box(self: Box<Self>) -> Result<Self::Writer, WriteError>

Equivalent to finish for boxed trait objects, because we otherwise can't move out of the unsized trait object.

Provided Methods

fn write_record(&mut self, bytes: &[u8]) -> Result<(), WriteError>