Enum telamon_utils::sequence::Sequence

pub enum Sequence<T> {
    List(List<T>),
    Vec(Vec<T>),
}

A type representing a sequence of values.

Can be implemented as either a persistent list or as a native vector. This is handy for serialization: we usually work with persistent lists but those share part of their structure and make them bad candidates for "direct" serialization. Instead, we serialize the persistent list as a sequence, and always deserialize it as a vector. By duplicating the objects when serializing we sidestep the problem of having pointers in the serialized stream.

Variants
List(List<T>)Vec(Vec<T>)
Methods

impl<T> Sequence<T>

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn to_vec(self) -> Vec<T> where
T: Clone,