Struct telamon_utils::VecSet

pub struct VecSet<T> { /* fields omitted */ }

A set backed by an ordered vector.

Methods

impl<T> VecSet<T> where
T: Ord,

pub fn new(data: Vec<T>) -> Self

Creates a new VecSet with the given data.

pub fn is_empty(&self) -> bool

Indicates if the VecSet is empty.

pub fn len(&self) -> usize

Returns the number of elements in the set.

pub fn iter(&self) -> Iter<T>

Iterates over the set, in order.

pub fn difference<'a>(&'a self, other: &'a VecSet<T>) -> Difference<T>

Returns the elements in self but not in other.

pub fn relative_difference(self, other: Self) -> (Self, Self)

Returns self\other and other\self.

pub fn intersection<'a>(&'a self, other: &'a VecSet<T>) -> Intersection<T>

Returns a set containing the elements present in both self and other.

pub fn intersect(&mut self, other: &VecSet<T>)

In-place intersection with another VecSet.

pub fn union(&self, other: &VecSet<T>) -> VecSet<T> where
T: Clone,

Returns a set containing the elements present in either selforother`.

pub fn filter<P>(&self, predicate: P) -> Self where
T: Clone,
P: FnMut(&T) -> bool,

Returns a new VecSet with only the elements for which the predicate returned true.

pub fn retain<P>(&mut self, predicate: P) where
P: FnMut(&T) -> bool,

Filters out elements for wich the predicate returns false.

pub fn insert(&mut self, item: T) -> bool

Inserts an element in the VecSet. This operation has a complexity in O(n). Returns false if the item was already present.

pub fn contains(&self, item: &T) -> bool

Indicates if the VecSet contains the given object.